Module breeze_core::dma

Implements DMA (Direct Memory Access).

The SNES provides 2 DMA mechanisms: General purpose DMA can be set up by the CPU by writing to $420B - MDMAEN, which stops the CPU and performs the transfer on all enabled channels. HDMA can be set up to perform one transfer per H-Blank per channel. There are 8 channels in total, and some of them may be configured to be used for HDMA, while the other channels can be used for DMA.

HDMA needs to be coordinated with the PPU: It needs to be initialized when a new frame starts, which is done by calling init_hdma, and can transfer data every scanline, which is done by do_hdma. DMA is simpler: It is started by calling do_dma when the CPU writes to $420B and doesn't need periodic callbacks.

Structs

DmaChannel
The configuration of a DMA channel

Enums

TransferMode
Describes how a single DMA unit (max. 4 Bytes) is transferred

Functions

do_dma
Performs all DMA transactions enabled by the given channels bitmask. Returns the number of master cycles spent.

do_hdma
Performs one H-Blank worth of HDMA transfers (at most 8, if all channels are enabled).

init_hdma
Refresh HDMA state for a new frame. This is called at V=0, H~6 and will set up some internal registers for all active channels.