Module breeze_core::ppu

Emulates the Picture Processing Unit.

We emulate the NTSC console, which has a refresh rate of 60 Hz and a screen resolution of 256x224 pixels by default.

Documentation mostly taken from http://emu-docs.org/Super%20NES/General/snesdoc.html and http://wiki.superfamicom.org/

Modules

cgram
CGRAM definitions and methods

oam
OAM (Object Attribute Memory) representation and utilities

Structs

FrameBuf

Ppu

Rgb
Standard 24-bit RGB color rendered to the frame buffer

SnesRgb
5-bit per channel RGB value used by the SNES

Vram

Constants

SCREEN_HEIGHT
Physical screen height (this is the height of a field, or a half-frame)

SCREEN_WIDTH
Physical screen width (this is the width of a field, or a half-frame)

VRAM_SIZE
VRAM size in Bytes