Struct breeze_core::ppu::cgram::Cgram

pub struct Cgram(_);

Methods

impl Cgram

fn get_color(&self, color: u8) -> SnesRgb

Looks up a color in CGRAM and returns the RGB color values stored inside, without adjusting the color range to full RGB.

fn get_color_raw(&self, color: u8) -> u16

Gets the raw, 16-bit (technically 15), color value stored at the given color index

Colors are stored as BGR big-endian values. This method returns the BGR data in host-endian order.

fn set_color_raw(&mut self, index: u8, raw: u16)

Set a raw color value

-bbbbbgg gggrrrrr

The raw color is converted to big endian and stored as 2 CGRAM bytes.