Struct breeze_core::ppu::Ppu

pub struct Ppu {
    pub framebuf: FrameBuf,
    pub oam: Oam,
    pub cgram: Cgram,
    pub vram: Vram,
    pub can_latch_counters: bool,
    // some fields omitted
}

Fields
framebuf: FrameBuf

PPU frame buffer. Contains raw RGB pixel data in RGB24 format: The first byte is the red component of the pixel in the top left corner of the frame, the second byte is the green component and the third byte is the blue component. The fourth byte is then the red component of the second pixel (at coordinate (1,0)), and so on.

oam: Oam

Object Attribute Memory

The first 512 Bytes contain 4 Bytes per sprite (for a maximum of 128 simultaneous on-screen sprites). The remaining 32 Bytes contain additional 2 bits per sprite.

Layout of an OAM entry (4 Bytes) in the first 512 Bytes: * xxxxxxxx - Low 8 bits of X coordinate of the sprite (top left corner?) * yyyyyyyy - Y coordinate * tttttttt - Starting tile/character number (low 8 bits) * vhoopppt - Vertical/**H**orizontal flip, Pri**o**rity bits, Palette number, Bit 9 (most significant bit) of starting tile number.

Layout of a byte in the last 32 Bytes of OAM: sxsxsxsx - Size toggle bit and most significant bit of X coordinate The low bits contain the information for sprites with low IDs.

cgram: Cgram

CGRAM - Stores the color palette

There are 256 colors in the palette, each 15 bits (5 bits per color channel), represented by 2 Bytes of CGRAM. Layout (16-bit big endian value - high byte and high address first): ?bbbbbgg gggrrrrr (the ?-bit is ignored)

vram: Vram

VRAM - Stores background maps and tile/character data

The location of background maps can be selected with the registers $2107-$210A. An entry in these maps looks like this:

vhopppcc cccccccc * Vertical/**H**orizontal flip * Pri**o**rity bit * Palette number * Character/Tile starting number

Character data locations are set with the registers $210B (BG1/2) and $210C (BG3/4).

can_latch_counters: bool

Set by the emulator on writes to $4201: When bit 7 of $4201 is 0, no latching can occur

Methods

impl Ppu

fn lookup_bg_color(&mut self, bg_num: u8, prio: u8, subscreen: bool) -> Option<SnesRgb>

Main entry point into the BG layer renderer.

Lookup the color of the given background layer (1-4) at the current pixel, using the given priority (0-1) only. This will also scroll backgrounds accordingly.

This may only be called with BG layer numbers which are actually valid in the current BG mode (the renderer code makes sure that this is the case).

Returns None if the pixel is transparent, Some(SnesRgb) otherwise.

impl Ppu

Rendering

fn obj_size(&self, size_toggle: bool) -> (u8, u8)

Get the configured sprite size in pixels. If size_toggle is false, gets the size of small sprites, otherwise gets the size of large sprites (OAM size bit set).

fn bg_mode(&self) -> u8

Returns the active BG mode (0-7).

fn render_pixel(&mut self) -> Rgb

Main rendering entry point. Renders the current pixel and returns its color. Assumes that the current pixel is on the screen.

fn read_chr_entry(&self, bitplane_count: u8, start_addr: u16, tile_size: u8, (x, y): (u8, u8), (vflip, hflip): (bool, bool)) -> u8

Reads character data for a pixel and returns the palette index stored in the bitplanes.

Parameters

bitplane_count: Number of bitplanes (must be even)

start_addr: Address of the first bitplane (or the first 2)

tile_size: 8 or 16

(x, y): Offset inside the tile (0-7 or 0-15, depending on the tile size)

(vflip, hflip): Flip this tile vertically (top and down are flipped) or horizontally (left and right are flipped)

impl Ppu

fn inidisp(&self) -> u8

fn obsel(&self) -> u8

fn bgmode(&self) -> u8

fn mosaic(&self) -> u8

fn bg1sc(&self) -> u8

fn bg2sc(&self) -> u8

fn bg3sc(&self) -> u8

fn bg4sc(&self) -> u8

fn bg12nba(&self) -> u8

fn bg34nba(&self) -> u8

fn vmain(&self) -> u8

fn m7sel(&self) -> u8

fn w12sel(&self) -> u8

fn w34sel(&self) -> u8

fn wobjsel(&self) -> u8

fn wh0(&self) -> u8

fn wh1(&self) -> u8

fn wh2(&self) -> u8

fn wh3(&self) -> u8

fn wbglog(&self) -> u8

fn wobjlog(&self) -> u8

fn tm(&self) -> u8

fn ts(&self) -> u8

fn tmw(&self) -> u8

fn tsw(&self) -> u8

fn cgwsel(&self) -> u8

fn cgadsub(&self) -> u8

fn coldata_r(&self) -> u8

fn coldata_g(&self) -> u8

fn coldata_b(&self) -> u8

fn setini(&self) -> u8

impl Ppu

fn bg1hofs(&self) -> u16

fn m7hofs(&self) -> u16

fn bg1vofs(&self) -> u16

fn m7vofs(&self) -> u16

fn bg2hofs(&self) -> u16

fn bg2vofs(&self) -> u16

fn bg3hofs(&self) -> u16

fn bg3vofs(&self) -> u16

fn bg4hofs(&self) -> u16

fn bg4vofs(&self) -> u16

fn m7a(&self) -> u16

fn m7b(&self) -> u16

fn m7c(&self) -> u16

fn m7d(&self) -> u16

fn m7x(&self) -> u16

fn m7y(&self) -> u16

impl Ppu

fn collect_sprite_data_for_scanline(&mut self)

Collects visible sprites and sprite tiles for the current scanline.

Called when rendering the first pixel on a scanline.

fn maybe_draw_sprite_pixel(&self, prio: u8, subscreen: bool) -> Option<(SnesRgb, bool)>

Returns the value of the current pixel on the sprite layer if it has the given priority (and None otherwise).

Returns the pixel's color and whether the sprite uses palette 0-3 (if this is the case, the sprite can not participate in color math - it is fixed to opaque).

impl Ppu

fn load(&mut self, addr: u16) -> u8

Load a PPU register (addresses $2134 to $213f)

fn store(&mut self, addr: u16, value: u8)

Store a byte in a PPU register (addresses $2100 - $2133)

fn latch_counters(&mut self)

Latches the H/V counters if $4201 bit 7 is set (otherwise, no latching can occur)

fn update(&mut self) -> u8

Runs the PPU for a bit.

This will render exactly one pixel (when in H/V-Blank, the pixel counter will be incremented, but obviously nothing will be drawn).

fn in_h_blank(&self) -> bool

fn in_v_blank(&self) -> bool

fn forced_blank(&self) -> bool

fn h_counter(&self) -> u16

Returns the current X position

fn v_counter(&self) -> u16

Returns the current Y position (scanline)