Struct breeze_core::snes::Emulator

pub struct Emulator<R: Renderer, A: AudioSink> {
    pub renderer: R,
    pub audio: A,
    pub snes: Snes,
    // some fields omitted
}

The emulator.

Fields
renderer: R

The renderer this emulator instance uses to display the screen

audio: A

The audio sink to be used for APU output

snes: Snes
Methods

impl<R: Renderer, A: AudioSink> Emulator<R, A>

fn new(rom: Rom, renderer: R, audio: A) -> Self

Creates a new emulator instance from a loaded ROM and a renderer.

This will also create a default Input instance without any attached peripherals.

fn peripherals(&self) -> &Peripherals

Get a reference to the Peripherals instance

fn peripherals_mut(&mut self) -> &mut Peripherals

Get a mutable reference to the Peripherals instance

fn handle_action(&mut self, action: BackendAction) -> bool

Handles a BackendAction. Returns true if the emulator should exit.

fn render_frame(&mut self) -> BackendResult<bool>

Runs emulation until a frame is completed, renders the frame and handles an action dictated by the backend.

Returns true if the backend requested an exit, false otherwise.

fn run(&mut self) -> BackendResult<()>

Runs the emulator in a loop

This will emulate the system and render frames until the backend signals that the emulator should exit.