Enum breeze_core::input::Peripheral

pub enum Peripheral {
    Joypad {
        imp: Box<JoypadImpl>,
        state: JoypadState,
    },
}

Enumeration of things that can be plugged into a controller port on the SNES.

Variants
Joypad

The standard SNES joypad: A, B, X, Y, L, R, Start, Select, D-Pad

Fields of Joypad

imp: Box<JoypadImpl>

The actual implementation (provided by the backend)

state: JoypadState

Current joypad state. When the latch is active, this is updated by asking the backend for the current state.

Methods

impl Peripheral

Construction

pub fn new_joypad(imp: Box<JoypadImpl>) -> Self

Creates a new joypad peripheral using the given JoypadImpl.

impl Peripheral

CPU interface

pub fn set_latch(&mut self, latch: bool)

Called when the value of the lowest bit of $4016 changes. When set to 1, the controller should latch its input (whatever that means is specific to the attached peripheral).

Auto-joypad mode writes 1 and then 0 to the latch before reading data.

pub fn read_bit(&mut self) -> (bool, bool)

Read a bit from the Data1 and Data2 lines. Called on serial reads either via Auto-Joypad mode or reads from $4016/$4017.

Returns the bits on Data1 and Data2, respectively. For Joypads, for example, Data2 will always be false, since it's not connected.

pub fn set_io_bit(&mut self, _iobit: bool)

Sets the bit written out to the IOBit line.

This is called when the SNES writes to the highest 2 bits of $4213. (If these are set to 0, reads from $4201 will always return 0. If these are set to 1, then reads from $4201 will return whatever value is written to the respective IOBit lines.)

pub fn read_io_bit(&mut self) -> bool

Called on reads from $4201 when the respective bit in $4213 is set to 1 (if the bit in $4213 is set to 0, all reads will return 0 and this method is not called).

This should return the current status of the IOBit line.

When using the IOBit line of port 2 to latch the PPU's H/V Counters, use needs_hv_latch_control and update_hv_latch in addition to this (the counters are latched when IOBit transitions from 1 to 0).

pub fn update_hv_latch(&mut self) -> bool

This will be called on every pixel. When this method returns true, the PPU's H/V Counters will be latched.

Note that the returned value is not returned on read from the I/O Port ($4201). You have to make sure that this method and read_io_bit return correct values.

pub fn next_frame(&mut self)

Called once after every frame