Struct breeze_core::input::Input

pub struct Input {
    pub ports: Ports,
    // some fields omitted
}

Controller input management.

Fields
ports: Ports
Methods

impl Input

pub fn start_recording(&mut self, rec: Box<Recorder>)

Start recording input to a Write implementor, often a file.

When reading data from a controller port, the recorder will write that data to the given Box<Write>.

pub fn start_replay(&mut self, replayer: Box<Replayer>)

Start replaying input from a recording made with start_recording. While replaying, user input is ignored (but input sources are still updated).

pub fn is_recording(&self) -> bool

pub fn is_replaying(&self) -> bool

pub fn new_frame(&mut self)

pub fn load(&mut self, reg: u16) -> u8

Read from an input register. Updates the controller state if this is the first load in this frame.

pub fn store(&mut self, reg: u16, val: u8)

Store to an input register. Stores to $4016 can change the latch line.

pub fn perform_auto_read(&mut self)

Called when auto joypad read is enabled and it's time to do one.

On the real console, auto joypad read takes place in the first few scanline in V-Blank. We pretend it's instantaneous and set the auto joypad read bit in $4212 manually.