Struct breeze_core::ppu::SnesRgb

pub struct SnesRgb { /* fields omitted */ }

5-bit per channel RGB value used by the SNES

Methods

impl SnesRgb

pub fn new(r: u8, g: u8, b: u8) -> SnesRgb

Creates a new SnesRgb instance from the given RGB values.

When debug assertions are enabled, this will panic when any color value is outside the 5 bit range. When assertions are disabled, the values will be truncated to their low 5 bits.

pub fn r(&self) -> u8

pub fn g(&self) -> u8

pub fn b(&self) -> u8

pub fn saturating_add(&self, other: &Self) -> Self

Performs saturating addition of self and other per color.

pub fn saturating_sub(&self, other: &Self) -> Self

Performs saturating subtraction of self and other per color.

pub fn to_adjusted_rgb(&self) -> Rgb

Converts 5-bit RGB to 8-bit RGB, adjusting the color space

The colors are adjusted as follows (http://wiki.superfamicom.org/snes/show/Palettes): text Rout = Rin << 3 Gout = Gin << 3 Bout = Bin << 3 Rout += Rout / 32 Gout += Gout / 32 Bout += Bout / 32